+++ src/game/themes/spring.ts
@@ -1,0 +1,39 @@
+import { BlockColor } from "../types";
+
+export const shapeMap: Record<BlockColor, string> = {
+  black: "🥚",
+  brown: "🐓",
+  darkgreen: "️🐥",
+  yellow: "🦢",
+  aqua: "️🌱",
+  pink: "🌷",
+  purple: "️️️🐇",
+  blue: "🍀",
+  red: "🪿",
+  white: "️️🦋",
+  green: "️🌳",
+  darkblue: "️️🎵",
+  turquoise: "🎀",
+  orange: "☀️",
+  lightyellow: "️️☁️",
+  gray: "️🧩",
+};
+
+export const colorMap: Record<BlockColor, string> = {
+  red: "#FFEB3B",
+  white: "#F06292",
+  yellow: "#42A5F5",
+  blue: "#81C784",
+  purple: "#FFAB91",
+  black: "#FFF8E1",
+  green: "#FFECB3",
+  darkgreen: "#AB47BC",
+  darkblue: "#E3F2FD",
+  aqua: "#64B5F6",
+  brown: "#4DD0E1",
+  pink: "#FF8A65",
+  turquoise: "#AED581",
+  orange: "#FFF59D",
+  lightyellow: "#FFB74D",
+  gray: "#26C6DA",
+};